Develop

+++ src/main/java/acme/features/developer/trainingModule/DeveloperTrainingModuleDeleteService.java
@@ -12,12 +12,15 @@
 
 package acme.features.developer.trainingModule;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import acme.client.data.models.Dataset;
 import acme.client.services.AbstractService;
 import acme.entities.training.TrainingModule;
+import acme.entities.training.TrainingSession;
 import acme.roles.Developer;
 
 @Service
@@ -34,12 +37,9 @@ public class DeveloperTrainingModuleDeleteService extends AbstractService<Develo
 	@Override
 	public void authorise() {
 		boolean status;
-		int trainingModuleId;
-		TrainingModule trainingModule;
-
-		trainingModuleId = super.getRequest().getData("id", int.class);
-		trainingModule = this.repository.findTrainingModuleById(trainingModuleId);
-		status = trainingModule != null;
+		int moduleId = super.getRequest().getData("id", int.class);
+		TrainingModule module = this.repository.findTrainingModuleById(moduleId);
+		status = module != null && module.isDraftMode() && super.getRequest().getPrincipal().hasRole(module.getDeveloper());
 
 		super.getResponse().setAuthorised(status);
 	}
@@ -68,19 +68,10 @@ public class DeveloperTrainingModuleDeleteService extends AbstractService<Develo
 	@Override
 	public void perform(final TrainingModule object) {
 		assert object != null;
+		List<TrainingSession> sessions = new ArrayList<>(this.repository.findTrainingSessionsOfTrainingModule(object.getId()));
+		if (sessions.isEmpty())
+			this.repository.delete(object);
 
-		this.repository.delete(object);
-	}
-
-	@Override
-	public void unbind(final TrainingModule object) {
-		assert object != null;
-
-		Dataset dataset;
-
-		dataset = super.unbind(object, "code", "creationMoment", "details", "difficulty", "updateMoment", "startTotalTime", "endTotalTime", "link", "draftMode");
-
-		super.getResponse().addData(dataset);
 	}
 
 }
